AAGIThemes: AAGI Branding for Graphical and Tabular Outputs

Applies 'AAGI' external brand guidelines to graphics. 'AAGI' colours and font guidelines are applied as useful and reasonable to base graphics, 'ggplot2' figures and 'flextable' objects.

See also

Author

Maintainer: Adam H. Sparks adam.sparks@dpird.wa.gov.au (ORCID)

Authors:

Kenyon Ng kenyon.ng@dpird.wa.gov.au (ORCID)

Tinula Kariyawasam tinula.kariyawasam@dpird.wa.gov.au

Matt Cowgill matthew.cowgill@grattaninstitute.edu.au (Wrote major portions of package for saving charts for 'Grattan' package, adapted to 'AAGIThemes' under 'MIT' licence.)

Will Mackey william.mackey@grattaninstitute.edu.au (Wrote major portions of package for saving charts for 'Grattan' package, adapted to 'AAGIThemes' under 'MIT' licence.)

Other contributors:

Jimmy Ng jimmy.ng@dpird.wa.gov.au [contributor]

Curtin University of Technology [copyright holder]